Extracted Quotation to its own module.
