+++ tests/test_quotes.py
@@ -13,7 +13,8 @@ This module is built on top of the Pydle system.
 import datetime
 from unittest import TestCase, expectedFailure
 
-from Modules.rat_rescue import Quotation, Rescue
+from Modules.rat_rescue import Rescue
+from Modules.rat_quotation import Quotation
 from Modules.trigger import Trigger
 from tests import mock_bot
